docker/taist_dev_api/requirements.txt
fastapi
uvicorn
pymongo
pydantic
mysql-connector-python